move settings determination to helper
