+++ package.json
@@ -61,7 +61,13 @@
     ],
     "rules": {
       "import/extensions": "off",
-      "import/no-extraneous-dependencies": "off"
+      "import/no-extraneous-dependencies": "off",
+      "@typescript-eslint/no-use-before-define": [
+        "error",
+        {
+          "functions": false
+        }
+      ]
     },
     "root": true
   },
